+++ sampledb/frontend/utils.py
@@ -229,12 +229,19 @@ def custom_format_date(date, format='%Y-%m-%d'):
 
 
 @jinja_filter('babel_format_number')
-def custom_format_number(number: typing.Union[str, int, float], display_digits: typing.Optional[int] = None) -> str:
+def custom_format_number(
+    number: typing.Union[str, int, float],
+    display_digits: typing.Optional[int] = None,
+    integral_digits: typing.Optional[int] = None,
+    disable_scientific_format: bool = False
+) -> str:
     """
     Return the formatted number.
 
     :param number: either an int or a float, or a string representation of either
     :param display_digits: number of decimals to use, or None
+    :param integral_digits: minimal number of digits in the integral part, or None
+    :param disable_scientific_format: disables output in scientific format
     :return: the formatted number
     """
     try:
@@ -253,15 +260,22 @@ def custom_format_number(number: typing.Union[str, int, float], display_digits: 
         number = round(number, display_digits)
 
     locale = get_locale()
-    if number == 0:
-        exponent = 0
+    if not disable_scientific_format:
+        if number == 0:
+            exponent = 0
+        else:
+            exponent = int(floor(log10(abs(number))))
+
+        # for very small or very large absolute numbers, the exponential format should be used
+        use_exponential_format = not -5 < exponent < 6
     else:
-        exponent = int(floor(log10(abs(number))))
+        exponent = 0
+        use_exponential_format = False
 
-    # for very small or very large absolute numbers, the exponential format should be used
-    use_exponential_format = not -5 < exponent < 6
-
-    format = None
+    if integral_digits is not None:
+        positive_format = '0' * integral_digits
+    else:
+        positive_format = '0'
     if display_digits is not None:
         if use_exponential_format:
             display_digits += exponent
@@ -271,18 +285,20 @@ def custom_format_number(number: typing.Union[str, int, float], display_digits: 
         if display_digits > 27:
             display_digits = 27
 
-        positive_format = '0.' + '0' * display_digits
-        if use_exponential_format:
-            # including E will enable exponential format, 0 means the exponent should be shown even if 0
-            positive_format += 'E0'
-        negative_format = '-' + positive_format
-        format = positive_format + ';' + negative_format
+        positive_format = positive_format + '.' + '0' * display_digits
+    else:
+        positive_format = positive_format + '.###'
+    if use_exponential_format:
+        # including E will enable exponential format, 0 means the exponent should be shown even if 0
+        positive_format += 'E0'
+    negative_format = '-' + positive_format
+    f = positive_format + ';' + negative_format
 
     if use_exponential_format:
         return numbers.format_scientific(
             number,
             locale=locale,
-            format=format,
+            format=f,
             decimal_quantization=False
         )
     else:
@@ -291,10 +307,37 @@ def custom_format_number(number: typing.Union[str, int, float], display_digits: 
             return numbers.format_decimal(
                 numbers.decimal.Decimal(number),
                 locale=locale,
-                format=format,
-                decimal_quantization=False,
-                group_separator=False
+                format=f,
+                decimal_quantization=False
             )
+
+
+@jinja_filter('format_time')
+def format_time(
+    magnitude_in_base_units: float,
+    units: str,
+    display_digits: typing.Optional[int] = None
+):
+    if units not in {'min', 'h'}:
+        raise errors.MismatchedUnitError()
+    decimal = 0
+    magnitude_str = str(magnitude_in_base_units)
+    if '.' in magnitude_str:
+        # string manipulation to prevent ugly float-arithmetics
+        integral_str, decimal_str = magnitude_str.split('.', maxsplit=1)
+        magnitude_in_base_units = int(integral_str)
+        decimal = float(f'0.{decimal_str}')
+    seconds = magnitude_in_base_units % 60
+    magnitude_in_base_units -= seconds
+    seconds += decimal
+    if units == 'h':
+        minutes = int(magnitude_in_base_units // 60) % 60
+        magnitude_in_base_units -= minutes * 60
+        hours = int(magnitude_in_base_units // 3600)
+        return f'{hours:02d}:{minutes:02d}:{custom_format_number(seconds, display_digits, 2, True)}'
+    if units == 'min':
+        minutes = int(magnitude_in_base_units // 60)
+        return f'{minutes:02d}:{custom_format_number(seconds, display_digits, 2, True)}'
 
 
 @jinja_filter('format_quantity')
@@ -307,8 +350,11 @@ def custom_format_quantity(
         return mdash
     if data.get('units', '1') in {'1', ''}:
         return custom_format_number(data.get('magnitude_in_base_units', 0), schema.get('display_digits', None))
-    quantity = Quantity.from_json(data)
     narrow_non_breaking_space = '\u202f'
+    magnitude = data.get('magnitude_in_base_units', 0)
+    if data.get('units') in {'h', 'min'}:
+        return f'{format_time(magnitude, data.get("units"), schema.get("display_digits"))}{narrow_non_breaking_space}{data.get("units")}'
+    quantity = Quantity.from_json(data)
     return custom_format_number(quantity.magnitude, schema.get('display_digits', None)) + narrow_non_breaking_space + prettify_units(quantity.units)